+++ service/bookturks/exceptions/NotImplementedException.py
@@ -1,0 +1,5 @@
+class NotImplementedException(Exception):
+    """
+    For not yet implemented code.
+    """
+    pass
