+++ service/bookturks/exceptions/__init__.py
@@ -1,0 +1,1 @@
+from service.bookturks.exceptions.NotImplementedException import NotImplementedException
