+++ service/bookturks/models/ControllerModel.py
@@ -1,0 +1,9 @@
+class ControllerModel(object):
+    """
+    Model for View.
+    """
+
+    def __init__(self, view, redirect, context=None):
+        self.view = view
+        self.redirect = redirect
+        self.context = context if context else dict()
