+++ service/bookturks/models/__init__.py
@@ -1,0 +1,6 @@
+from service.bookturks.models.ControllerModel import ControllerModel
+from service.bookturks.models.EventModel import EventModel
+from service.bookturks.models.NotificationModel import NotificationModel
+from service.bookturks.models.QuizCompleteModel import QuizCompleteModel
+from service.bookturks.models.QuizResultModel import QuizResultModel
+from service.bookturks.models.UserProfileModel import UserProfileModel
