+++ service/bookturks/storage_handlers/UserProfileStorageHandler.py
@@ -1,0 +1,7 @@
+class UserProfileStorageHandler(object):
+    """
+    Storage handler for User Profiles
+    """
+
+    def __init__(self):
+        pass
