+++ service/bookturks/user_views/user_home_view.py
@@ -1,17 +1,15 @@
 """
 User Home View
 """
-from django.core.urlresolvers import reverse
-from django.http import HttpResponseRedirect
-from django.shortcuts import render
 
-from service.bookturks.Constants import USER_HOME_PAGE, REQUEST, USER, ALERT_MESSAGE, ALERT_TYPE, SERVICE_USER_SETUP, \
-    USER_PROFILE_MODEL
-from service.bookturks.adapters.UserAdapter import UserAdapter
-from service.bookturks.alerts import init_alerts
-from service.bookturks.user.UserProfileTools import UserProfileTools
+from service.bookturks.Constants import USER_HOME_PAGE, SERVICE_USER_SETUP, USER_PROFILE_MODEL
+from service.bookturks.adapters import UserAdapter
+from service.bookturks.decorators.Controller import controller
+from service.bookturks.models import ControllerModel
+from service.bookturks.utils.UserProfileTools import UserProfileTools
 
 
+@controller
 def user_home_main_view(request):
     """
     User Home page. Lands on the dashboard page
@@ -19,23 +17,17 @@ def user_home_main_view(request):
     :param request:
     :return: Renders user_home page (dashboard)
     """
-    request, alert_type, alert_message = init_alerts(request=request)
-
     user_adapter = UserAdapter()
     user_profile_tools = UserProfileTools()
 
     if USER_PROFILE_MODEL not in request.session:
         user_model = user_adapter.get_user_instance_from_django_user(request.user)
         if not user_model:
-            return HttpResponseRedirect(reverse(SERVICE_USER_SETUP))
+            return ControllerModel(view=SERVICE_USER_SETUP, redirect=True)
         request.session[USER_PROFILE_MODEL] = user_profile_tools.get_profile(user_model)
 
     context = {
-        REQUEST: request,
-        USER: request.user,
-        ALERT_MESSAGE: alert_message,
-        ALERT_TYPE: alert_type,
         USER_PROFILE_MODEL: request.session.get(USER_PROFILE_MODEL)
     }
 
-    return render(request, USER_HOME_PAGE, context)
+    return ControllerModel(view=USER_HOME_PAGE, redirect=False, context=context)
