+++ service/bookturks/user_views/user_myquiz_view.py
@@ -1,42 +1,35 @@
 """
 My quizzes page
 """
-from django.core.urlresolvers import reverse
-from django.http import HttpResponseRedirect
-from django.shortcuts import render
 
-from service.bookturks.Constants import REQUEST, USER, ALERT_MESSAGE, ALERT_TYPE, DANGER, \
-    USER_MYQUIZ_HOME_PAGE, USER_MYQUIZ_INFO_PAGE, SERVICE_USER_HOME, SERVICE_USER_MYQUIZ_HOME
-from service.bookturks.adapters.QuizAdapter import QuizAdapter
-from service.bookturks.adapters.UserAdapter import UserAdapter
-from service.bookturks.alerts import init_alerts, set_alert_session
-from service.bookturks.quiz.QuizTools import QuizTools
+from service.bookturks.Constants import DANGER, \
+    USER_MYQUIZ_HOME_PAGE, USER_MYQUIZ_INFO_PAGE, SERVICE_USER_MYQUIZ_HOME, SERVICE_USER_HOME
+from service.bookturks.adapters import QuizAdapter, UserAdapter
+from service.bookturks.alerts import set_alert_session
+from service.bookturks.decorators.Controller import controller
+from service.bookturks.models import ControllerModel
+from service.bookturks.storage_handlers import QuizStorageHandler
 
 
+@controller
 def user_myquiz_home_view(request):
     """
     My quizzes home page
     :param request: User request
     :return: Renders page
     """
-    request, alert_type, alert_message = init_alerts(request=request)
     user_adapter = UserAdapter()
     quiz_adapter = QuizAdapter()
 
     user = user_adapter.get_user_instance_from_django_user(request.user)
 
     quiz_list = quiz_adapter.get_models_for_owner(user_model=user)
-    context = {
-        REQUEST: request,
-        USER: request.user,
-        ALERT_MESSAGE: alert_message,
-        ALERT_TYPE: alert_type,
-        'quiz_list': quiz_list
-    }
+    context = {'quiz_list': quiz_list}
 
-    return render(request, USER_MYQUIZ_HOME_PAGE, context)
+    return ControllerModel(view=USER_MYQUIZ_HOME_PAGE, redirect=False, context=context)
 
 
+@controller
 def user_myquiz_info_view(request, quiz_id):
     """
     Quiz related tasks
@@ -46,30 +39,39 @@ def user_myquiz_info_view(request, quiz_id):
     """
     user_adapter = UserAdapter()
     quiz_adapter = QuizAdapter()
-    quiz_tools = QuizTools()
+    quiz_storage_handler = QuizStorageHandler()
 
-    user = user_adapter.get_user_instance_from_django_user(request.user)
-    quiz = quiz_adapter.exists(quiz_id)
-    if not quiz or not user:
+    try:
+        user = user_adapter.get_user_instance_from_django_user(request.user)
+        quiz = quiz_adapter.exists(quiz_id)
+        if not quiz or not user:
+            raise ValueError("No such quiz exists")
+
+        # Check if this user is the owner of this quiz
+        if user != quiz.quiz_owner:
+            raise ValueError("You do not own this quiz. You are not allowed to edit this")
+
+    except ValueError as err:
         set_alert_session(session=request.session,
-                          message="No such quiz exists",
+                          message=str(err),
                           alert_type=DANGER)
-        return HttpResponseRedirect(reverse(SERVICE_USER_HOME))
+        return ControllerModel(view=SERVICE_USER_HOME, redirect=True)
 
-    # Check if this user is the owner of this quiz
-    if user != quiz.quiz_owner:
-        set_alert_session(session=request.session, message="You do not own this quiz. You are not allowed to edit this",
+    # Different try block because redirection is different.
+    try:
+        # Downloading the content from storage.
+        content = quiz_storage_handler.download_quiz_content(quiz_model=quiz)
+        if not content:
+            raise ValueError("This quiz is unavailable")
+
+    except ValueError as err:
+        set_alert_session(session=request.session,
+                          message=str(err),
                           alert_type=DANGER)
-        return HttpResponseRedirect(reverse(SERVICE_USER_HOME))
-
-    # Downloading the content from storage.
-    content = quiz_tools.download_quiz_content(quiz_model=quiz)
-    if not content:
-        set_alert_session(session=request.session, message="This quiz is unavailable", alert_type=DANGER)
-        return HttpResponseRedirect(reverse(SERVICE_USER_MYQUIZ_HOME))
+        return ControllerModel(view=SERVICE_USER_MYQUIZ_HOME, redirect=True)
 
     # Passing these session objects as this page submits the result to user_quiz_verifier view which checks for them.
     request.session['quiz_complete_model'] = content
 
     context = {'quiz_data': content.quiz_data}
-    return render(request, USER_MYQUIZ_INFO_PAGE, context)
+    return ControllerModel(view=USER_MYQUIZ_INFO_PAGE, redirect=False, context=context)
