+++ service/tests/test_unittest/test_user_quizarena_view.py
@@ -79,7 +79,7 @@ class UserQuizArenaViewTest(QuizTest):
         session.save()
 
         # Preparing mock file for test
-        with open("".join([settings.BASE_DIR, "/service/tmp", self.quiz_tools.create_filename(quiz)]),
+        with open("".join([settings.BASE_DIR, "/service/tmp", self.quiz_storage_handler.create_filename(quiz)]),
                   'wb') as mock_file:
             mock_file.write(serialize(quiz_complete_model))
             mock_file.close()
@@ -147,7 +147,7 @@ class UserQuizArenaViewTest(QuizTest):
         session.save()
 
         # Preparing mock file for test
-        with open("".join([settings.BASE_DIR, "/service/tmp", self.quiz_tools.create_filename(quiz)]),
+        with open("".join([settings.BASE_DIR, "/service/tmp", self.quiz_storage_handler.create_filename(quiz)]),
                   'wb') as mock_file:
             mock_file.write(serialize(quiz_complete_model))
             mock_file.close()
